refactor: localstorage -> localStorage 폴더 네임 변경
